feat: finalizados os testes unitários para o service de InformativoTipo. Todos os testes estão funcionando. Adicionei validação para o método excluir do service verificar se o id existe antes de apagar o registro

+++ src/main/java/com/project/ecopraia/service/InformativoTipoService.java
@@ -46,6 +46,10 @@ public class InformativoTipoService {
     }
 
     public void excluir(Long id){
+        if(!informativoTipoRepository.existsById(id)) {
+            throw  new RuntimeException("Informativo não encontrado");
+        }
+        
         informativoTipoRepository.deleteById(id);
     }
 }
